requirements.txt
python-crontab==3.2.0
sqlalchemy==2.0.0
pandas==2.2.0
requests==2.32.0
beautifulsoup4==4.12.0
google-auth==2.36.0
google-api-python-client==2.152.0
pytest-cov==6.0.0